HACK: cpuidle-debugfs-tegra23x: Do not access cnthp_ctl_el2 with pkvm
If we are in kvm protected mode, we are not allowed to access EL2 registers
and thus we cannot access the cnthp_ctl_el2. As a workaround we return
-EBUSY to the cpudile framework.

This is a nasty workaround because it will prevent the CPU to enter idle
state. A better solution is needed but I will use this for testing
purposes.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>