drivers/video/tegra/dc/ext/dev.c: Fix ifdef statement
This function should not be included unless CONFIG_TEGRA_ISOMGR is also
defined. We must fix this to be able to build without CONFIG_TEGRA_ISOMGR
because the nvidia tree does ignorantly build everything in this
directory with -Werror, so when this warns for an unused function, it turns
into an error that stops compilation.

This is another fix to make the nvidia kernel sources compile for use
with QEMU.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@unikie.com>